Update link rel values to reflect commons changes
