+++ src/community/ogcapi/ogcapi-core/src/main/java/org/geoserver/ogcapi/AbstractLandingPageDocument.java
@@ -5,56 +5,16 @@
 package org.geoserver.ogcapi;
 
 import com.fasterxml.jackson.annotation.JsonPropertyOrder;
-import io.swagger.v3.oas.models.OpenAPI;
-import java.util.function.BiConsumer;
-import org.springframework.http.MediaType;
 
 /**
  * A class representing a generic API service landing page in a way that Jackson can easily
  * translate to JSON/YAML (and can be used as a Freemarker template model)
  */
 @JsonPropertyOrder({"title", "description", "links"})
-public class AbstractLandingPageDocument extends AbstractDocument {
-
-    final String title;
-    final String description;
+public class AbstractLandingPageDocument extends AbstractLandingPageDocumentNoConformance {
 
     public AbstractLandingPageDocument(String title, String description, String serviceBase) {
-        final APIRequestInfo requestInfo = APIRequestInfo.get();
-
-        // self and alternate representations of landing page
-        addLinksFor(
-                serviceBase + "/",
-                AbstractLandingPageDocument.class,
-                "This document as ",
-                "landingPage",
-                new BiConsumer<MediaType, Link>() {
-                    boolean first = true;
-
-                    @Override
-                    public void accept(MediaType mediaType, Link link) {
-                        if ((first
-                                && requestInfo.isFormatRequested(
-                                        mediaType, MediaType.APPLICATION_JSON))) {
-                            link.setRel(Link.REL_SELF);
-                            link.setTitle("This document");
-                            first = false;
-                        }
-                    }
-                },
-                Link.REL_ALTERNATE);
-        // api
-        addLinksFor(
-                serviceBase + "/api",
-                OpenAPI.class,
-                "API definition for this endpoint as ",
-                "api",
-                (format, link) -> {
-                    if (MediaType.TEXT_HTML.equals(format)) {
-                        link.setRel(Link.REL_SERVICE_DOC);
-                    }
-                },
-                Link.REL_SERVICE_DESC);
+        super(title, description, serviceBase);
         // conformance
         addLinksFor(
                 serviceBase + "/conformance",
@@ -62,16 +22,6 @@ public class AbstractLandingPageDocument extends AbstractDocument {
                 "Conformance declaration as ",
                 "conformance",
                 null,
-                Link.REL_CONFORMANCE);
-        this.title = title;
-        this.description = description;
-    }
-
-    public String getTitle() {
-        return title;
-    }
-
-    public String getDescription() {
-        return description;
+                Link.REL_CONFORMANCE_URI);
     }
 }
